requirements.txt
pandas
tabula-py
numpy